+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+## Added
+- Thrayir, Eyes of the Siren
+- Cyrce's Circlet & Citrines
+
 ## [1.4.4] 2024-12-19
 ## Changes
 - Bump TOC to 11.0.5, 11.0.7